Add activity indicator when searching for mythbackend
When we're searching for a mythbackend provide some kind
of feedback to the user via an activity indicator that a search
is in progress. After 20 seconds of not finding a mythbackend
then provide an error message that we weren't able find a
mythbackend. This is using SwiftSpinner from icanzilb on
GitHub.